+++ src/connector/__init__.py
@@ -1,5 +1,5 @@
 """Init file."""
 
-from connector.main import create_engine, create_movement_plan
+from connector.main import create_engine
 
-__all__ = ["create_engine", "create_movement_plan"]
+__all__ = ["create_engine"]